Added delay to resetting vertical inputs after they are released. Makes it easier to fire vertically and diagonally. ContactSensor detects when contact state changes, so that other components can take action if required. MovementActuator limits upwards speed when contact state changes to wallgrab, to prevent sliding up walls when jumping to early and double jumping up into a wall.

+++ 2-5d-character-controller/Assets/Scripts/Abstract Input/AbstractInput.cs
@@ -24,10 +24,10 @@ public abstract class AbstractInput : MonoBehaviour{
 	//To use a virtual controller, must call UpdateInput() prior To
 	//getting any axis/button states.
 	public abstract void UpdateInput();
-	public float GetHorAxis(){
+	public int GetHorAxis(){
 		return horAxis;
 	}
-	public float GetVertAxis(){
+	public int GetVertAxis(){
 		return vertAxis;
 	}
 	public bool GetJumpDown(){
